nvm: port posix interface to work under env_nvm
This is to allow specific file types to accessed through the posix interface.
This is a momentary solution. The idea is refactore the posix backed so that it
is available to other backeds too.

nvm: use possix for information log
nvm: port rest of posix functions to nvm environment
nvm: fix directory creation for posix when dir exists
nvm: return children from both posix and dflash