Enable storage backend to store private metadata
Summary: This patch provides a set of helper functions that would enable
a storage backend (e.g., posix, HDFS) to store sstable private metadata
in the manifest. This frees the storage backend from implementing its
own metadata mechanisms by reusing the functionality already present in
the manifest. This same idea can be applied to the WAL and MANIFEST; if
the mechanism is correct, patches for these will follow. We follow the
new MANIFEST format, which enables to define custom tags in a new file
type (kNewFile4).

Test Plan: This patch has been tested in posix and in a new backend for
OpenChannel SSDs [1], where the private metadata functions have been
implemented. The idea is that testing functions are provided by the
storage backend when used.

[1] https://github.com/OpenChannelSSD