Added Flush(bool) to signal whether the last buffer should be flushed to media.
Somewhere in rocks, WritableFileWriter, writes are buffered to satisfy
alignment requirements by env. This has the effect that writes and
done using PositionedAppend instead of Append.

PositionedAppend is not used for entirely random writes, it is just used
to enable re-writing padded data in the buffer. The unfortunate side-effect
is that byte-offset are written more than once, which will require a flash-erase
to satisfy.

This change uses a hack that relaxes the Flush, it simply does not flash the
current last buffer. The buffer will be flashed when:

1) Flush occurs and that buffer is no longer the last :)
2) File is closed

This is done to avoid re-writing the same page